+++ src/Application/Cart/PaymentMethod/PaymentMethodForCartRepository.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Trader\Application\Cart\PaymentMethod;
+
+interface PaymentMethodForCartRepository
+{
+    /** @return PaymentMethodForCart[] */
+    public function findAllPaymentMethodsForCart(?string $countryId = null): array;
+}