+++ tests/TestHelpers.php
@@ -35,6 +35,7 @@ use Thinktomorrow\Trader\Domain\Model\Order\Shipping\ShippingState;
 use Thinktomorrow\Trader\Domain\Model\Order\Shopper;
 use Thinktomorrow\Trader\Domain\Model\Order\State\OrderState;
 use Thinktomorrow\Trader\Domain\Model\PaymentMethod\PaymentMethod;
+use Thinktomorrow\Trader\Domain\Model\PaymentMethod\PaymentMethodState;
 use Thinktomorrow\Trader\Domain\Model\Product\Option\Option;
 use Thinktomorrow\Trader\Domain\Model\Product\Option\OptionId;
 use Thinktomorrow\Trader\Domain\Model\Product\Option\OptionValue;
@@ -297,9 +298,12 @@ trait TestHelpers
     {
         return PaymentMethod::fromMappedData(array_merge([
             'payment_method_id' => 'mmm',
+            'state' => PaymentMethodState::online->value,
             'rate' => '123',
             'data' => json_encode([]),
-        ], $values));
+        ], $values), [
+            CountryId::class => [],
+        ]);
     }
 
     protected function createShippingProfile(array $values = []): ShippingProfile